requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-jose==3.3.0
sqlalchemy==2.0.36
